+++ changelog.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 2021-01-12
+
+- Expansion of `.` and `~` in volume configurations is now also supported if the config is all on a single line. (e.g.: `-v=~/exchange`). Close #1
+- Expansion of `.` and `~` in volume configurations is now also supported if the config is espressed with the `--mount` format (see <https://docs.docker.com/storage/bind-mounts/>)
+
 ## 2021-01-11 
 
 - Modified sample `dockerstarter.yaml`